Move Spring Scheduling RunnableWrapper to separate class
This works around an edge case where Spring Sleuth tries to load the RunnableWrapper's enclosing class:
https://github.com/spring-projects/spring-framework/blob/v5.2.2.RELEASE/spring-core/src/main/java/org/springframework/core/annotation/AnnotationsScanner.java#L233

Since the enclosing class is on a separate class loader, this fails and results in an exception.
